LayoutTests/printing/css2.1/page-break-before-000.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN">
<html>
<head>
<meta http-equiv="Content-Type" content="text/html;charset=UTF-8">
<title>CSS Test: page-break-before: always</title>
<link rel="author" href="mailto:mturnwall@revenution.com" title="Michael Turnwall">
<link rel="author" href="http://www.hp.com/" title="Hewlett-Packard Company">
<link rel="help" href="http://www.w3.org/TR/CSS21/page.html#page-break-props">
<link rel="help" href="http://www.w3.org/TR/css3-page/#pg-br-before-after">
<meta content="paged" name="flags">
<meta content="The 'always' value of the 'page-break-before' property forces a page break before the box." name="assert">
<style type="text/css">

    div.break {
        page-break-before:always;
    }

</style>
<script src="../resources/paged-media-test-utils.js"></script>
<script>

    if (window.testRunner)
        testRunner.dumpAsText();

    function test()
    {
        pageNumberForElementShouldBe('test1', 0);
        pageNumberForElementShouldBe('test2', 1);
        numberOfPagesShouldBe(2);
    }

</script>
</head>
<body onload="runPrintingTest(test)">
    <div id="test1">This test should produce two pages of output. This text should be at the top of the first page.</div>
    <div class="break" id="test2">This text should be at the top of the second and final page.</div>
</body>
</html>
